% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeAndCheckInput.R
\name{checkValidInput}
\alias{checkValidInput}
\title{Checks if input to modCluster is valid. If valid returns igraph object as
described by igraph::isgraph}
\usage{
checkValidInput(edges, vert)
}
\arguments{
\item{edges}{valid edges as described by igraph::graph_from_data_frame}

\item{vert}{valid verticies as described by igraph::graph_from_data_frame}
}
\value{
graphOject when valid, error when invalid
}
\description{
Checks if input to modCluster is valid. If valid returns igraph object as
described by igraph::isgraph
}
