% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseGroups.R
\name{chooseBestGroups}
\alias{chooseBestGroups}
\title{Selects best groups by highest density metrics}
\usage{
chooseBestGroups(sorted, minModSize)
}
\arguments{
\item{sorted}{sorted output of makeSortedCompare}

\item{minModSize}{int size of the smallest module}
}
\value{
string list of best groups in sorted
}
\description{
Selects best groups by highest density metrics
}
