% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeAndCheckInput.R
\name{clusterByModule}
\alias{clusterByModule}
\title{Controller function available to the user to display}
\usage{
clusterByModule(edges, vert, displayCommunity = TRUE)
}
\arguments{
\item{edges}{valid edges as described by igraph::graph_from_data_frame}

\item{vert}{valid verticies as described by igraph::graph_from_data_frame}

\item{displayCommunity}{a boolean flag to choose to display community markers
(community markers is TRUE, no community markers is FALSE)}
}
\description{
Controller function available to the user to display
}
\examples{
\dontrun{
load("./data/exampleEdge.Rda")
load("./data/exampleVert.Rda")
clusterByModule(edges, vert, displayCommunity=FALSE)
clusterByModule(edges, vert, displayCommunity=TRUE)
}
}
