% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_density.R
\name{edge_density}
\alias{edge_density}
\title{Modified igraph::edge_density to include weights}
\usage{
edge_density(graph)
}
\arguments{
\item{graph}{a valid igraph object}
}
\value{
float representing density of the graph
}
\description{
Modified igraph::edge_density to include weights
}
