% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edges}
\alias{edges}
\title{Example data of edges in the sample graph}
\format{A data frame with 3 rows and 15 variables:
\describe{
  \item{GENE1}{the first gene in the edge}
  \item{GENE2}{the second gene in the edge}
  \item{WEIGHT}{the weight of the edge}
}}
\source{
\url{https://www.wormbase.org/}
}
\usage{
edges(...)
}
\description{
A dataset containing hypothetical edges and verticies using C. elegans genes
}
