% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgesData.R
\docType{data}
\name{edgesData}
\alias{edgesData}
\title{Example data of edges in the sample graph}
\format{A data frame with 3 rows and 15 variables:
\describe{
  \item{GENE1}{gene in edge}
  \item{GENE2}{gene in edge}
  \item{WEIGHT}{the weight of a given edge}
}}
\source{
\url{https://www.wormbase.org/}
}
\usage{
edgesData
}
\description{
A dataset containing hypothetical edges using C. elegans genes
}
\keyword{datasets}
