% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllPossibleGroups.R
\name{getAllPossibleGroups}
\alias{getAllPossibleGroups}
\title{Makes df containing all possible combined modules of length 2 with density calculated}
\usage{
getAllPossibleGroups(graph, vert, moduleGraphs)
}
\arguments{
\item{graph}{a valid graph object}

\item{vert}{valid verticies as described by igraph::graph_from_data_frame}

\item{moduleGraphs}{df with $MODULE module name, $SUBGRAPH graph for verts in module, $DENSITY density of subgraph}
}
\value{
df like moduleGraphs, where each module is every possible module combination of size 2
}
\description{
Makes df containing all possible combined modules of length 2 with density calculated
}
