% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayoutAndChosen.R
\name{getChosenMods}
\alias{getChosenMods}
\title{Get df with module name, subgraph and density for the groups with the best density}
\usage{
getChosenMods(bestGroups, moduleGraphs, grouped)
}
\arguments{
\item{bestGroups}{strings of best combined module groups}

\item{moduleGraphs}{df with $MODULE, $SUBGRAPH, $DENSITY for provided modules}

\item{grouped}{df with $MODULE, $SUBGRAPH, $DENSITY for all grouped modules}
}
\value{
df with $MODULE, $SUBGRAPH, $DENSITY for the modules in bestGroups
}
\description{
Get df with module name, subgraph and density for the groups with the best density
}
