% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleNames.R
\name{getModuleNames}
\alias{getModuleNames}
\title{Helper function that converts modules from string to vector}
\usage{
getModuleNames(mods)
}
\arguments{
\item{mods}{string representing modules}
}
\value{
vector of modules
}
\description{
Helper function that converts modules from string to vector
}
