% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayoutAndChosen.R
\name{makeLayout}
\alias{makeLayout}
\title{Use igraph::mergecoords to make a layout with the chosen modules}
\usage{
makeLayout(selected)
}
\arguments{
\item{selected}{df with module name, subgraph and density for the groups with the best density}
}
\value{
layout, df with xy coordinates
}
\description{
Use igraph::mergecoords to make a layout with the chosen modules
}
