% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModuleGraphs.R
\name{makeModuleGraphs}
\alias{makeModuleGraphs}
\title{Make df with module name, subgraph and density}
\usage{
makeModuleGraphs(edges, vert)
}
\arguments{
\item{edges}{valid edges as described by igraph::graph_from_data_frame}

\item{vert}{valid verticies as described by igraph::graph_from_data_frame}
}
\value{
df with columns (MODULES, SUBGRAPHS, DENSITY) from modules in verts
}
\description{
Make df with module name, subgraph and density
}
