% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{makeSortedCompare}
\alias{makeSortedCompare}
\title{Helper function to modCluster
Combines df with modules and combined modules and sorts in descending}
\usage{
makeSortedCompare(ungrouped, grouped)
}
\arguments{
\item{ungrouped}{df with $MODULE ungrouped module name $SUBGRAPH igraph object $DENSITY density of subgraph}

\item{grouped}{df with $MODULE grouped module name $SUBGRAPH igraph object $DENSITY density of subgraph}
}
\value{
sorted df
}
\description{
Helper function to modCluster
Combines df with modules and combined modules and sorts in descending
}
