% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modCluster.R
\name{modCluster}
\alias{modCluster}
\title{Controller function available to the user to display}
\usage{
modCluster(edges, vert, displayGene = FALSE)
}
\arguments{
\item{edges}{valid edges as described by igraph::graph_from_data_frame}

\item{vert}{valid verticies as described by igraph::graph_from_data_frame}

\item{displayGene}{a flag to choose to display or hide genes}
}
\description{
Controller function available to the user to display
}
