% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modClusterShiny.R
\name{modClusterShiny}
\alias{modClusterShiny}
\title{Run shiny app that shows and visualizes example}
\usage{
modClusterShiny()
}
\description{
Run shiny app that shows and visualizes example
}
\examples{
\dontrun{
modCluster::runExample()
}
}
