% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticiesData.R
\docType{data}
\name{verticiesData}
\alias{verticiesData}
\title{Example data of verticies in the sample graph}
\format{A data frame with 2 rows and 11 variables:
\describe{
  \item{GENE}{the gene}
  \item{MODULE}{the module the gene belongs to}
}}
\source{
\url{https://www.wormbase.org/}
}
\usage{
verticiesData
}
\description{
A dataset containing hypothetical verticies using C. elegans genes
}
\keyword{datasets}
