% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeCommunities.R
\name{visualizeCommunities}
\alias{visualizeCommunities}
\title{Visualizes graph communities}
\usage{
visualizeCommunities(graph, chosen, layout, membership, communities,
  communitiesList)
}
\arguments{
\item{graph}{a valid igraph object}

\item{chosen}{df with chosen grouped modules}

\item{layout}{xy layout object}

\item{membership}{list of gene names and module name}

\item{communities}{igraph communities object}

\item{communitiesList}{list of modules}
}
\description{
Visualizes graph communities
}
