% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeAndCheckInput.R
\name{visualizeGraph}
\alias{visualizeGraph}
\title{Visualizes graph with gene names}
\usage{
visualizeGraph(graph, chosen, layout, membership, communitiesList)
}
\arguments{
\item{graph}{a valid igraph object}

\item{chosen}{df with chosen grouped modules}

\item{layout}{xy layout object}

\item{membership}{list of gene names and module name}

\item{communitiesList}{list of modules}
}
\value{
0
}
\description{
Visualizes graph with gene names
}
